+++ astropy/modeling/core.py
@@ -34,20 +34,21 @@ from ..utils import indent, isinstancemethod, metadata
 from ..extern import six
 from ..extern.six.moves import copyreg, zip
 from ..table import Table
+from ..units import Quantity, UnitBase, UnitsError, dimensionless_unscaled
+from ..units.utils import quantity_asanyarray
 from ..utils import (sharedmethod, find_current_module,
-                     check_broadcast, IncompatibleShapeError,
-                     InheritDocstrings, OrderedDescriptorContainer)
+                     InheritDocstrings, OrderedDescriptorContainer,
+                     check_broadcast, IncompatibleShapeError, isiterable)
 from ..utils.codegen import make_function_with_signature
 from ..utils.compat import suppress
 from ..utils.compat.funcsigs import signature
 from ..utils.exceptions import AstropyDeprecationWarning
-from .utils import (array_repr_oneline, combine_labels,
-                    make_binary_operator_eval, ExpressionTree,
-                    AliasDict, get_inputs_and_params,
-                    _BoundingBox)
+from .utils import (combine_labels, make_binary_operator_eval,
+                    ExpressionTree, AliasDict, get_inputs_and_params,
+                    _BoundingBox, _combine_equivalency_dict)
 from ..nddata.utils import add_array, extract_array
 
-from .parameters import Parameter, InputParameterError
+from .parameters import Parameter, InputParameterError, param_repr_oneline
 
 
 __all__ = ['Model', 'FittableModel', 'Fittable1DModel', 'Fittable2DModel',
@@ -358,6 +359,7 @@ class _ModelMeta(OrderedDescriptorContainer, InheritDocstrings, abc.ABCMeta):
                     {'__call__': __call__})
 
     def _handle_special_methods(cls, members):
+
         # Handle init creation from inputs
         def update_wrapper(wrapper, cls):
             # Set up the new __call__'s metadata attributes as though it were
@@ -372,31 +374,55 @@ class _ModelMeta(OrderedDescriptorContainer, InheritDocstrings, abc.ABCMeta):
 
         if ('__call__' not in members and 'inputs' in members and
                 isinstance(members['inputs'], tuple)):
-            inputs = members['inputs']
-            # Done create a custom __call__ for classes that already have one
+
+            # Don't create a custom __call__ for classes that already have one
             # explicitly defined (this includes the Model base class, and any
             # other classes that manually override __call__
+
             def __call__(self, *inputs, **kwargs):
                 """Evaluate this model on the supplied inputs."""
-
                 return super(cls, self).__call__(*inputs, **kwargs)
 
+            # When called, models can take two optional keyword arguments:
+            #
+            # * model_set_axis, which indicates (for multi-dimensional input)
+            #   which axis is used to indicate different models
+            #
+            # * equivalencies, a dictionary of equivalencies to be applied to
+            #   the input values, where each key should correspond to one of
+            #   the inputs.
+            #
+            # The following code creates the __call__ function with these
+            # two keyword arguments.
+            inputs = members['inputs']
             args = ('self',) + inputs
             new_call = make_function_with_signature(
-                __call__, args, [('model_set_axis', None), ('with_bounding_box', False), ('fill_value', np.nan)])
+                    __call__, args, [('model_set_axis', None),
+                                     ('with_bounding_box', False),
+                                     ('fill_value', np.nan),
+                                     ('equivalencies', None)])
+
+            # The following makes it look like __call__ was defined in the class
             update_wrapper(new_call, cls)
+
             cls.__call__ = new_call
 
         if ('__init__' not in members and not inspect.isabstract(cls) and
                 cls._parameters_):
+
             # If *all* the parameters have default values we can make them
-            # keyword arguments; otherwise they must all be positional
-            # arguments
-            if all(p.default is not None
-                   for p in six.itervalues(cls._parameters_)):
+            # keyword arguments; otherwise they must all be positional arguments
+            if all(p.default is not None for p in six.itervalues(cls._parameters_)):
                 args = ('self',)
-                kwargs = [(name, cls._parameters_[name].default)
-                          for name in cls.param_names]
+                kwargs = []
+                for param_name in cls.param_names:
+                    default = cls._parameters_[param_name].default
+                    unit = cls._parameters_[param_name].unit
+                    # If the unit was specified in the parameter but the default
+                    # is not a Quantity, attach the unit to the default.
+                    if unit is not None:
+                        default = Quantity(default, unit, copy=False)
+                    kwargs.append((param_name, default))
             else:
                 args = ('self',) + cls.param_names
                 kwargs = {}
@@ -652,6 +678,22 @@ class Model(object):
     # model hasn't completed initialization yet
     _n_models = 1
 
+    # Enforce strict units on inputs to evaluate. If this is set to True, input
+    # values to evaluate have to be in the exact right units specified by
+    # input_units. In this case, if the input quantities are convertible to
+    # input_units, they are converted.
+    input_units_strict = False
+
+    # Allow dimensionless input (and corresponding output). If this is True,
+    # input values to evaluate will gain the units specified in input_units.
+    # Only has an effect if input_units is defined.
+    input_units_allow_dimensionless = False
+
+    # Default equivalencies to apply to input values. If set, this should be a
+    # dictionary where each key is a string that corresponds to one of the model
+    # inputs. Only has an effect if input_units is defined.
+    input_units_equivalencies = None
+
     def __init__(self, *args, **kwargs):
         super(Model, self).__init__()
         meta = kwargs.pop('meta', None)
@@ -680,8 +722,13 @@ class Model(object):
         Evaluate this model using the given input(s) and the parameter values
         that were specified when the model was instantiated.
         """
+
         inputs, format_info = self.prepare_inputs(*inputs, **kwargs)
-        parameters = self._param_sets(raw=True)
+
+        # Check whether any of the inputs are quantities
+        inputs_are_quantity = any([isinstance(i, Quantity) for i in inputs])
+
+        parameters = self._param_sets(raw=True, units=True)
         with_bbox = kwargs.pop('with_bounding_box', False)
         fill_value = kwargs.pop('fill_value', np.nan)
         bbox = None
@@ -741,7 +788,24 @@ class Model(object):
         if self.n_outputs == 1:
             outputs = (outputs,)
 
-        return self.prepare_outputs(format_info, *outputs, **kwargs)
+        outputs = self.prepare_outputs(format_info, *outputs, **kwargs)
+
+        # If input values were quantities, we use return_units to cast
+        # the return values to the units specified by return_units.
+        if self.return_units and inputs_are_quantity:
+            # We allow a non-iterable unit only if there is one output
+            if self.n_outputs == 1 and not isiterable(self.return_units):
+                return_units = {self.outputs[0]: self.return_units}
+            else:
+                return_units = self.return_units
+
+            outputs = tuple([Quantity(out, return_units[out_name], subok=True)
+                             for out, out_name in zip(outputs, self.outputs)])
+
+        if self.n_outputs == 1:
+            return outputs[0]
+        else:
+            return outputs
 
     # *** Arithmetic operators for creating compound models ***
     __add__ =     _model_oper('+')
@@ -1094,6 +1158,108 @@ class Model(object):
 
     # *** Public methods ***
 
+    def without_units_for_data(self, **kwargs):
+        """
+        Return an instance of the model for which the parameter values have been
+        converted to the right units for the data, then the units have been
+        stripped away.
+
+        The input and output Quantity objects should be given as keyword
+        arguments.
+
+        Notes
+        -----
+
+        This method is needed in order to be able to fit models with units in
+        the parameters, since we need to temporarily strip away the units from
+        the model during the fitting (which might be done by e.g. scipy
+        functions).
+
+        The units that the parameters should be converted to are not necessarily
+        the units of the input data, but are derived from them. Model subclasses
+        that want fitting to work in the presence of quantities need to define a
+        _parameter_units_for_data_units method that takes the input and output
+        units (as two dictionaries) and returns a dictionary giving the target
+        units for each parameter.
+        """
+
+        model = self.copy()
+
+        inputs_unit = {inp: getattr(kwargs[inp], 'unit', dimensionless_unscaled)
+                       for inp in self.inputs if kwargs[inp] is not None}
+
+        outputs_unit = {out: getattr(kwargs[out], 'unit', dimensionless_unscaled)
+                        for out in self.outputs if kwargs[out] is not None}
+
+        parameter_units = self._parameter_units_for_data_units(inputs_unit, outputs_unit)
+
+        for name, unit in parameter_units.items():
+            parameter = getattr(model, name)
+            if parameter.unit is not None:
+                parameter.value = parameter.quantity.to(unit).value
+                parameter._set_unit(None, force=True)
+
+        return model
+
+    def with_units_from_data(self, **kwargs):
+        """
+        Return an instance of the model which has units for which the parameter
+        values are compatible with the data units specified.
+
+        The input and output Quantity objects should be given as keyword
+        arguments.
+
+        Notes
+        -----
+
+        This method is needed in order to be able to fit models with units in
+        the parameters, since we need to temporarily strip away the units from
+        the model during the fitting (which might be done by e.g. scipy
+        functions).
+
+        The units that the parameters will gain are not necessarily the units of
+        the input data, but are derived from them. Model subclasses that want
+        fitting to work in the presence of quantities need to define a
+        _parameter_units_for_data_units method that takes the input and output
+        units (as two dictionaries) and returns a dictionary giving the target
+        units for each parameter.
+        """
+
+        model = self.copy()
+
+        inputs_unit = {inp: getattr(kwargs[inp], 'unit', dimensionless_unscaled)
+                       for inp in self.inputs if kwargs[inp] is not None}
+
+        outputs_unit = {out: getattr(kwargs[out], 'unit', dimensionless_unscaled)
+                        for out in self.outputs if kwargs[out] is not None}
+
+        parameter_units = self._parameter_units_for_data_units(inputs_unit, outputs_unit)
+
+        # We are adding units to parameters that already have a value, but we
+        # don't want to convert the parameter, just add the unit directly, hence
+        # the call to _set_unit.
+        for name, unit in parameter_units.items():
+            parameter = getattr(model, name)
+            parameter._set_unit(unit, force=True)
+
+        return model
+
+    @property
+    def _has_units(self):
+        # Returns True if any of the parameters have units
+        for param in self.param_names:
+            if getattr(self, param).unit is not None:
+                return True
+        else:
+            return False
+
+    @property
+    def _supports_unit_fitting(self):
+        # If the model has a '_parameter_units_for_data_units' method, this
+        # indicates that we have enough information to strip the units away
+        # and add them back after fitting, when fitting quantities
+        return hasattr(self, '_parameter_units_for_data_units')
+
     @abc.abstractmethod
     def evaluate(self, *args, **kwargs):
         """Evaluate the model on some input variables."""
@@ -1213,15 +1379,72 @@ class Model(object):
 
         return out
 
+    @property
+    def input_units(self):
+        """
+        This property is used to indicate what units or sets of units the
+        evaluate method expects, and returns a dictionary mapping inputs to
+        units (or `None` if any units are accepted).
+
+        Model sub-classes can also use function annotations in evaluate to
+        indicate valid input units, in which case this property should
+        not be overriden since it will return the input units based on the
+        annotations.
+        """
+        if hasattr(self, '_input_units'):
+            return self._input_units
+        elif hasattr(self.evaluate, '__annotations__'):
+            annotations = self.evaluate.__annotations__.copy()
+            annotations.pop('return', None)
+            if annotations:
+                # If there are not annotations for all inputs this will error.
+                return dict((name, annotations[name]) for name in self.inputs)
+        else:
+            # None means any unit is accepted
+            return None
+
+    @input_units.setter
+    def input_units(self, input_units):
+        self._input_units = input_units
+
+    @property
+    def return_units(self):
+        """
+        This property is used to indicate what units or sets of units the output
+        of evaluate should be in, and returns a dictionary mapping outputs to
+        units (or `None` if any units are accepted).
+
+        Model sub-classes can also use function annotations in evaluate to
+        indicate valid output units, in which case this property should not be
+        overriden since it will return the return units based on the
+        annotations.
+        """
+        if hasattr(self, '_return_units'):
+            return self._return_units
+        elif hasattr(self.evaluate, '__annotations__'):
+            return self.evaluate.__annotations__.get('return', None)
+        else:
+            # None means any unit is accepted
+            return None
+
+    @return_units.setter
+    def return_units(self, return_units):
+        self._return_units = return_units
+
     def prepare_inputs(self, *inputs, **kwargs):
         """
         This method is used in `~astropy.modeling.Model.__call__` to ensure
         that all the inputs to the model can be broadcast into compatible
         shapes (if one or both of them are input as arrays), particularly if
-        there are more than one parameter sets.
+        there are more than one parameter sets. This also makes sure that (if
+        applicable) the units of the input will be compatible with the evaluate
+        method.
         """
 
+        # When we instantiate the model class, we make sure that __call__ can
+        # take the following two keyword arguments.
         model_set_axis = kwargs.pop('model_set_axis', None)
+        equivalencies = kwargs.pop('equivalencies', None)
 
         if model_set_axis is None:
             # By default the model_set_axis for the input is assumed to be the
@@ -1236,6 +1459,80 @@ class Model(object):
 
         _validate_input_shapes(inputs, self.inputs, n_models,
                                model_set_axis, self.standard_broadcasting)
+
+        # Check that the units are correct, if applicable
+
+        if self.input_units is not None:
+
+            # We combine any instance-level input equivalencies with user
+            # specified ones at call-time.
+            input_units_equivalencies = _combine_equivalency_dict(self.inputs,
+                                                                 equivalencies,
+                                                                 self.input_units_equivalencies)
+
+            # We now iterate over the different inputs and make sure that their
+            # units are consistent with those specified in input_units.
+            for i in range(len(inputs)):
+
+                input_name = self.inputs[i]
+
+                if self.input_units is not None:
+                    input_unit = self.input_units.get(input_name, None)
+
+                if isinstance(inputs[i], Quantity):
+
+                    # We check for consistency of the units with input_units,
+                    # taking into account any equivalencies
+
+                    if input_unit is None:
+
+                        # We dont need to do anything since no restrictions were
+                        # placed on input_units for this parameter.
+                        pass
+
+                    elif inputs[i].unit.is_equivalent(input_unit, equivalencies=input_units_equivalencies[input_name]):
+
+                        # If equivalencies have been specified, we need to
+                        # convert the input to the input units - this is because
+                        # some equivalencies are non-linear, and we need to be
+                        # sure that we evaluate the model in its own frame
+                        # of reference. If input_units_strict is set, we also
+                        # need to convert to the input units.
+                        if len(input_units_equivalencies) > 0 or self.input_units_strict:
+                            inputs[i] = inputs[i].to(input_unit, equivalencies=input_units_equivalencies[input_name])
+
+                    else:
+
+                        # We consider the following two cases separately so as
+                        # to be able to raise more appropriate/nicer exceptions
+
+                        if input_unit is dimensionless_unscaled:
+                            raise UnitsError("Units of input '{0}', {1} ({2}), could not be "
+                                             "converted to required dimensionless "
+                                             "input".format(self.inputs[i],
+                                                            inputs[i].unit,
+                                                            inputs[i].unit.physical_type))
+                        else:
+                            raise UnitsError("Units of input '{0}', {1} ({2}), could not be "
+                                             "converted to required input units of "
+                                             "{3} ({4})".format(self.inputs[i],
+                                                                inputs[i].unit,
+                                                                inputs[i].unit.physical_type,
+                                                                input_unit,
+                                                                input_unit.physical_type))
+                else:
+
+                    # If we allow dimensionless input, we add the units to the
+                    # input values without conversion, otherwise we raise an
+                    # exception.
+
+                    if (not self.input_units_allow_dimensionless and
+                       input_unit is not dimensionless_unscaled and input_unit is not None):
+                        if np.any(inputs[i] != 0):
+                            raise UnitsError("Units of input '{0}', (dimensionless), could not be "
+                                             "converted to required input units of "
+                                             "{1} ({2})".format(self.inputs[i], input_unit,
+                                                                input_unit.physical_type))
 
         # The input formatting required for single models versus a multiple
         # model set are different enough that they've been split into separate
@@ -1411,11 +1708,17 @@ class Model(object):
                 "given)".format(self.__class__.__name__, len(self.param_names),
                                 len(args)))
 
+        self._model_set_axis = model_set_axis
+        self._param_metrics = defaultdict(dict)
+
         for idx, arg in enumerate(args):
             if arg is None:
                 # A value of None implies using the default value, if exists
                 continue
-            params[self.param_names[idx]] = np.asanyarray(arg, dtype=np.float)
+            # We use quantity_asanyarray here instead of np.asanyarray because
+            # if any of the arguments are quantities, we need to return a
+            # Quantity object not a plain Numpy array.
+            params[self.param_names[idx]] = quantity_asanyarray(arg, dtype=np.float)
 
         # At this point the only remaining keyword arguments should be
         # parameter names; any others are in error.
@@ -1428,7 +1731,10 @@ class Model(object):
                 value = kwargs.pop(param_name)
                 if value is None:
                     continue
-                params[param_name] = np.asanyarray(value, dtype=np.float)
+                # We use quantity_asanyarray here instead of np.asanyarray because
+                # if any of the arguments are quantities, we need to return a
+                # Quantity object not a plain Numpy array.
+                params[param_name] = quantity_asanyarray(value, dtype=np.float)
 
         if kwargs:
             # If any keyword arguments were left over at this point they are
@@ -1439,9 +1745,6 @@ class Model(object):
                 raise TypeError(
                     '{0}.__init__() got an unrecognized parameter '
                     '{1!r}'.format(self.__class__.__name__, kwarg))
-
-        self._model_set_axis = model_set_axis
-        self._param_metrics = defaultdict(dict)
 
         # Determine the number of model sets: If the model_set_axis is
         # None then there is just one parameter set; otherwise it is determined
@@ -1494,8 +1797,10 @@ class Model(object):
         total_size = 0
 
         for name in self.param_names:
+            param_descr = getattr(self, name)
+
             if params.get(name) is None:
-                default = getattr(self, name).default
+                default = param_descr.default
 
                 if default is None:
                     # No value was supplied for the parameter and the
@@ -1506,8 +1811,13 @@ class Model(object):
                         "{1!r}".format(self.__class__.__name__, name))
 
                 value = params[name] = default
+                unit = param_descr.unit
             else:
                 value = params[name]
+                if isinstance(value, Quantity):
+                    unit = value.unit
+                else:
+                    unit = None
 
             param_size = np.size(value)
             param_shape = np.shape(value)
@@ -1516,6 +1826,14 @@ class Model(object):
 
             param_metrics[name]['slice'] = param_slice
             param_metrics[name]['shape'] = param_shape
+
+            if unit is None and param_descr.unit is not None:
+                raise InputParameterError(
+                    "{0}.__init__() requires a Quantity for parameter "
+                    "{1!r}".format(self.__class__.__name__, name))
+
+            param_metrics[name]['orig_unit'] = unit
+
             total_size += param_size
 
         self._param_metrics = param_metrics
@@ -1611,7 +1929,7 @@ class Model(object):
                 "to the broadcasting rules.".format(param_a, shape_a,
                                                     param_b, shape_b))
 
-    def _param_sets(self, raw=False):
+    def _param_sets(self, raw=False, units=False):
         """
         Implementation of the Model.param_sets property.
 
@@ -1627,26 +1945,33 @@ class Model(object):
 
         param_metrics = self._param_metrics
         values = []
+        shapes = []
         for name in self.param_names:
+            param = getattr(self, name)
+
             if raw:
-                value = getattr(self, name)._raw_value
+                value = param._raw_value
             else:
-                value = getattr(self, name).value
+                value = param.value
 
             broadcast_shape = param_metrics[name].get('broadcast_shape')
             if broadcast_shape is not None:
                 value = value.reshape(broadcast_shape)
 
+            shapes.append(np.shape(value))
+
+            if len(self) == 1:
+                # Add a single param set axis to the parameter's value (thus
+                # converting scalars to shape (1,) array values) for
+                # consistenc
+                value = np.array([value])
+
+            if units and param.unit is not None:
+                value = Quantity(value, param.unit)
+
             values.append(value)
 
-        shapes = [np.shape(value) for value in values]
-
-        if len(self) == 1:
-            # Add a single param set axis to the parameter's value (thus
-            # converting scalars to shape (1,) array values) for consistency
-            values = [np.array([value]) for value in values]
-
-        if len(set(shapes)) != 1:
+        if len(set(shapes)) != 1 or units:
             # If the parameters are not all the same shape, converting to an
             # array is going to produce an object array
             # However the way Numpy creates object arrays is tricky in that it
@@ -1658,6 +1983,11 @@ class Model(object):
             psets = np.empty(len(values), dtype=object)
             psets[:] = values
             return psets
+
+        # TODO: Returning an array from this method may be entirely pointless
+        # for internal use--perhaps only the external param_sets method should
+        # return an array (and just for backwards compat--I would prefer to
+        # maybe deprecate that method)
 
         return np.array(values)
 
@@ -1676,7 +2006,7 @@ class Model(object):
 
         parts.extend(
             "{0}={1}".format(name,
-                             array_repr_oneline(getattr(self, name).value))
+                             param_repr_oneline(getattr(self, name)))
             for name in self.param_names)
 
         if self.name is not None:
@@ -1724,6 +2054,9 @@ class Model(object):
 
         if columns:
             param_table = Table(columns, names=self.param_names)
+            # Set units on the columns
+            for name in self.param_names:
+                param_table[name].unit = getattr(self, name).unit
             parts.append(indent(str(param_table), width=4))
 
         return '\n'.join(parts)
@@ -2227,6 +2560,7 @@ class _CompoundModelMeta(_ModelMeta):
             # a bound Parameter even if submodel is a Model instance (as
             # opposed to a Model subclass)
             new_param = orig_param.copy(name=param_name, default=default,
+                                        unit=orig_param.unit,
                                         **constraints)
 
             setattr(cls, param_name, new_param)
@@ -2542,6 +2876,14 @@ class _CompoundModel(Model):
     def _get_submodels(self):
         return self.__class__._get_submodels()
 
+    def _parameter_units_for_data_units(self, input_units, output_units):
+        units_for_data = {}
+        for imodel, model in enumerate(self._submodels):
+            units_for_data_sub = model._parameter_units_for_data_units(input_units, output_units)
+            for param_sub in units_for_data_sub:
+                param = self._param_map_inverse[(imodel, param_sub)]
+                units_for_data[param] = units_for_data_sub[param_sub]
+        return units_for_data
 
 def custom_model(*args, **kwargs):
     """
@@ -2849,10 +3191,7 @@ def _prepare_outputs_single_model(model, outputs, format_info):
             else:
                 outputs[idx] = output.reshape(broadcast_shape)
 
-    if model.n_outputs == 1:
-        return outputs[0]
-    else:
-        return tuple(outputs)
+    return tuple(outputs)
 
 
 def _prepare_inputs_model_set(model, params, inputs, n_models, model_set_axis,
@@ -2931,10 +3270,7 @@ def _prepare_outputs_model_set(model, outputs, format_info):
             outputs[idx] = np.rollaxis(output, pivot,
                                        model.model_set_axis)
 
-    if model.n_outputs == 1:
-        return outputs[0]
-    else:
-        return tuple(outputs)
+    return tuple(outputs)
 
 
 def _validate_input_shapes(inputs, argnames, n_models, model_set_axis,
@@ -2991,7 +3327,6 @@ def _validate_input_shapes(inputs, argnames, n_models, model_set_axis,
                 arg_a, shape_a, arg_b, shape_b))
 
     return input_broadcast
-
 
 copyreg.pickle(_ModelMeta, _ModelMeta.__reduce__)
 copyreg.pickle(_CompoundModelMeta, _CompoundModelMeta.__reduce__)