+++ astropy/modeling/utils.py
@@ -12,11 +12,12 @@ from collections import deque, MutableMapping
 import numpy as np
 
 from ..extern import six
-from ..extern.six.moves import range, zip_longest, zip
+from ..extern.six.moves import range, zip
 
 from ..utils import isiterable, check_broadcast
 from ..utils.compat.funcsigs import signature
 
+from .. import units as u
 
 __all__ = ['ExpressionTree', 'AliasDict', 'check_broadcast',
            'poly_map_domain', 'comb', 'ellipse_extent']
@@ -520,12 +521,13 @@ def ellipse_extent(a, b, theta):
 
     Parameters
     ----------
-    a : float
+    a : float or `~astropy.units.Quantity`
         Major axis.
-    b : float
+    b : float or `~astropy.units.Quantity`
         Minor axis.
-    theta : float
-        Rotation angle in radians.
+    theta : float or `~astropy.units.Quantity`
+        Rotation angle. If given as a floating-point value, it is assumed to be
+        in radians.
 
     Returns
     -------
@@ -571,7 +573,10 @@ def ellipse_extent(a, b, theta):
     t = np.arctan2(b, a * np.tan(theta))
     dy = b * np.sin(t) * np.cos(theta) + a * np.cos(t) * np.sin(theta)
 
-    return np.abs([dx, dy])
+    if isinstance(dx, u.Quantity) or isinstance(dy, u.Quantity):
+        return np.abs(u.Quantity([dx, dy]))
+    else:
+        return np.abs([dx, dy])
 
 
 def get_inputs_and_params(func):
@@ -601,3 +606,31 @@ def get_inputs_and_params(func):
             params.append(param)
 
     return inputs, params
+
+
+def _parameter_with_unit(parameter, unit):
+    if parameter.unit is None:
+        return parameter.value * unit
+    else:
+        return parameter.quantity.to(unit)
+
+
+def _parameter_without_unit(value, old_unit, new_unit):
+    if old_unit is None:
+        return value
+    else:
+        return value * old_unit.to(new_unit)
+
+
+def _combine_equivalency_dict(keys, eq1=None, eq2=None):
+    # Given two dictionaries that give equivalencies for a set of keys, for
+    # example input value names, return a dictionary that includes all the
+    # equivalencies
+    eq = {}
+    for key in keys:
+        eq[key] = []
+        if eq1 is not None and key in eq1:
+            eq[key].extend(eq1[key])
+        if eq2 is not None and key in eq2:
+            eq[key].extend(eq2[key])
+    return eq